google_auth/requirements.txt
google-auth-oauthlib==1.0.0
google-api-python-client==2.86.0
PyJWT==2.7.0
FastAPI==0.100.0
Jinja2==3.1.2
python-dotenv==1.0.0
uvicorn==0.22.0